Extract compact_binary in separate crate to avoid cyclic dependencies
