+++ crates/compact-binary-derive/Cargo.toml
@@ -9,3 +9,4 @@ proc-macro = true
 [dependencies]
 quote = "1.0.37"
 syn = "2.0.90"
+stwo-compact-binary = { path = "../compact-binary" }
