+++ crates/stwo/Cargo.toml
@@ -50,10 +50,9 @@ rayon = { workspace = true, optional = true }
 serde.workspace = true
 tracing-subscriber.workspace = true
 hashbrown.workspace = true
-stwo-compact-binary-derive = { path = "../compact-binary-derive" }
 std-shims.workspace = true
-unsigned-varint.workspace = true
-zip = "4.3.0"
+stwo-compact-binary = { path = "../compact-binary" }
+stwo-compact-binary-derive = { path = "../compact-binary-derive" }
 
 [dev-dependencies]
 aligned = "0.4.2"
